Fix linear filtering on flats in software mode
The test for falling back to point filtering used the wrong variable to
determine if a span was too far away to bother with linear.

This test was (xfrac>mag_threshold && yfrac>mag_threshold). However
(xfrac,yfrac) is a projection of screen pixels to map coordinates, and,
in particular, depends on where the player is standing.

This is consistent with the observation first made by entryway that the
filtering stopped working when the player is standing to the right of
the map's y-axis.

Instead we must use the variables (xstep,ystep). When drawing a span
these define the displacement in map coordinates of one screen pixel to
the next. These displacements increase with distance from the view
point, and I believe the original intention was to use them to determine
when to fall back to point filtering. Certainly the default value of
mag_threshold gives a sensible maximum draw distance before fallback
when tested against (xstep, ystep).

Furthermore we must test that their absolute value is larger than
mag_threshold, or the fallback to point filtering will occur only when
the player is looking in certain directions.

(Incidentally the reason flats filtering worked in PrBoom 2.3 was
because the test was bracketed with "#if R_DRAWCOLUMN", instead of
R_DRAWSPAN as it should have been. Thus the fallback test was never even
compiled, and linear filtering was done everywhere, even at extremely
long draw distances where it made no sense to do it)